Initial commit
This is a very simple Golang application to look for K8s pods that are
using pullPolicy: Always and that are currently running an outdated
container image. This commit adds the main Golang app, the container
image and other bits needed to deploy it to Kubernetes will be added in
a follow-up.

Bug: T314705
Change-Id: I44bfef03eca5942a0b637a6494358d97478f45dd